meta-qoriq: replace MACHINE with SDK_MACHINE
Issue: LINUXEXEC-27724

Because the MACHINE value in WRLinux is different the one in NXP sdk
project, define variable SDK_MACHINE, and set it as the sdk MACHINE value
by BSP machine conf file. For example, it can be ls1028ardb, ls1043ardb,
ls1046ardb and so on. Finally, repace MACHINE with SDK_MACHINE so that
the correct parameter is passed.

(LOCAL REV: NOT UPSTREAM) - WRLinux specific

Signed-off-by: Meng Li <Meng.Li@windriver.com>