conf/layer.conf: remove problematic reset
Issue: LIN1024-2027

OE sets the default for a reason! And it's expected that all initramfs
images set IMAGE_NAME_SUFFIX to empty. This has been done in both OE
and wrlinux. Such reset is problematic. In fact, settings such as this
should never appear in a layer.conf file, because it will affect the
whole project. Remove this problematic setting. I cannot find any
reference in this layer. But whatever the initial problem is (the original
commit does not say it, and the issue ID looks like an internal issue ID),
it needs to be solved in a different way.

The original reset is there to fix some NXP's internal issue, so this
patch is marked as 'WRLinux Specific'.

(LOCAL REV: NOT UPSTREAM) -- WRLinux Specific

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>